+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Features
 
   * Added `suggestedMethod` method to `Route` class.
+  * Added `formTag` helper.
 
 ## v1.1.0 (2022-06-24)